Huge Update, Class seperation

+++ Setting.py
@@ -1,28 +1,32 @@
 import datetime
 import time
 
-import numpy
+import numpy as np
 import torch
 import torchvision
 from torchvision import datasets
 
-from dlg import attack
+from Dlg import Dlg
+from Prediction import Predictor
 from net import Net1, Net2, weights_init
 from test import test
 from train import train
-from prediction import predict_setting
 
 
 class Setting:
     def __init__(self, **kwargs):
-        self.check_cuda()
-
-        self.result = None
+        self.initialised = False
         self.target = []
         self.device = None
+        self.check_cuda()
         self.train_dataset = None
         self.test_dataset = None
         self.model = None
+
+        self.dlg = None
+        self.ids = []
+        self.orig_data = None
+        self.orig_label = None
 
         self.parameter = {}
         self.restore_default_parameter()
@@ -32,10 +36,21 @@ class Setting:
         self.load_dataset()
         self.load_model()
 
+        self.result = None
+        self.dlg = Dlg(self)
+        self.predictor = Predictor(self)
+        self.initialised = True
 
     def configure(self, **kwargs):
         for key, value in kwargs.items():
-            self.parameter[key] = value
+            if key == "target":
+                self.target = value
+                self.fix_ids()
+            elif key == "ids":
+                self.ids = value
+                self.fix_targets()
+            else:
+                self.parameter[key] = value
 
             # The following parameter changes require additional action
             if key == "dataset":
@@ -46,8 +61,9 @@ class Setting:
                 self.reset_seeds()
             elif key == "max_epoch_size" and value == 0:
                 self.parameter["max_epoch_size"] = len(self.train_dataset) / self.parameter["batch_size"]
-            elif key == "target":
-                self.target = value
+
+        # Renew Attack with new settings, does not execute yet
+        if self.initialised: self.dlg = Dlg(self)
 
     def restore_default_parameter(self):
         self.parameter = {
@@ -78,7 +94,7 @@ class Setting:
         # Check CUDA
         if torch.cuda.is_available():
             self.device = torch.device("cuda")
-            print("Torch successfully connected to CUDA device: {}".format(torch.cuda.current_device()))
+            #print("Torch successfully connected to CUDA device: {}".format(torch.cuda.current_device()))
         else:
             print("Error: Torch can't connect to CUDA")
             exit()
@@ -87,10 +103,10 @@ class Setting:
         # Setting Torch and numpy Seed
         if self.parameter["use_seed"]:
             torch.manual_seed(self.parameter["seed"])
-            numpy.random.seed(self.parameter["seed"])
+            np.random.seed(self.parameter["seed"])
         else:
             torch.manual_seed(int(1000 * time.time() % 2 ** 32))
-            numpy.random.seed(int(1000 * time.time() % 2 ** 32))
+            np.random.seed(int(1000 * time.time() % 2 ** 32))
 
     def load_dataset(self):
         # Initialising datasets
@@ -116,7 +132,7 @@ class Setting:
             exit()
 
     def load_model(self):
-        # prepare the model
+        # prepare the model, reloads it, undoes training
         if self.parameter["model"] == 1:
             self.model = Net1(self.parameter)
             self.model.apply(weights_init)
@@ -135,25 +151,68 @@ class Setting:
         test(self)
 
     def attack(self):
-        self.result = attack(self)
+        self.dlg.attack()
 
     def store_everything(self):
+        if self.result is None: exit("No result found, run attack first")
         self.result.store_everything()
 
     def store_composed_image(self):
+        if self.result is None: exit("No result found, run attack first")
         self.result.store_composed_image()
 
     def store_separate_images(self):
+        if self.result is None: exit("No result found, run attack first")
         self.result.store_separate_images()
 
     def store_data(self):
+        if self.result is None: exit("No result found, run attack first")
         self.result.store_data()
 
     def show_composed_image(self):
+        if self.result is None: exit("No result found, run attack first")
         self.result.show_composed_image()
 
     def delete(self):
+        if self.result is None: exit("No result found, run attack first")
         self.result.delete()
 
-    def predict(self):
-        return predict_setting(self)
+    def predict(self, *verbose):
+        self.predictor = Predictor(self)
+        self.predictor.predict()
+        if verbose:
+            orig = self.target[:self.parameter["batch_size"]].copy()
+            orig.sort()
+            #print("Orig:", orig)
+            #print("Pred:", self.predictor.prediction)
+            print(
+            "Correct: {}\tFalse: {}\tAcc: {}\tUsing: {}".format(self.predictor.correct, self.predictor.false,
+                                                                self.predictor.acc, self.parameter["prediction"]))
+        return self.predictor.prediction
+
+    def fix_targets(self):
+        self.target = []
+        for i in range(len(self.ids)):
+            self.target.append(self.train_dataset[self.ids[i]][1])
+
+    def fix_ids(self):
+        self.ids = []
+        for i in range(len(self.target)):
+            # searching for a sample with target label
+            for i_s, sample in enumerate(self.train_dataset):
+                # does this sample have the right label and was not used before?
+                if sample[1] == self.target[i] and not self.ids.__contains__(i_s):
+                    self.ids.append(i_s)
+                    break
+
+    def fill_targets(self):
+        # fill missing targets if underspecified
+        for i in range(self.parameter["batch_size"] - len(self.target)):
+            self.target.append(np.random.randint(0, self.parameter["num_classes"]))
+
+        self.fix_ids()
+
+    def copy(self):
+        return Setting(**self.parameter)
+
+
